Flooring error resolved
There was a flooring error.
Let there are 7 tasks t be divided into 3 persons.
they must be 2+2+3 (as much as possible equal)
But was assigning as 2+2+2+1
here 1 is assigned to 4th person which even doesn't exist.

So the check was made to verify no. of tasks per person (round off) is less then no of tasks (exact value).
Then it means flooring occurred.
So on this special case append the last person tasks instead of appending tasks for one person.